Add files via upload
This project is a Parking Management System that provides efficient management and tracking of vehicles in a parking area.  It offers a range of functionalities, including adding new vehicle entries, removing existing entries, viewing the list of parked vehicles with their details, and displaying the available parking space for each vehicle type,generating bills.
The system maintains a database of vehicles, including their vehicle number, type (Bicycle, Bike, Car), vehicle name, owner name, check-in date, and check-in time. Users can input vehicle details and the system validates the information to ensure accuracy.

The system also provides information on parking rates for different vehicle types, allowing users to determine the charges based on the duration of the vehicle's stay. Additionally, it calculates additional charges at a rate of 18% for the parking fee and presents the total amount due.

This project demonstrates effective utilization of programming concepts such as data structures (lists) for storing vehicle information, input validation, looping for continuous operation, and exception handling to ensure smooth functioning of the system.